+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Fixed issue where parallel assignment statements and assignment to multiple targets were not checked by `redundant_assignment_checker`
 - Fixed issue where annotated assignment statements were not checked by `redundant_assignment_checker`
 - Fixed issue where empty preconditions were preventing CFGs from being generated
+- Added strict numeric type checking to enforce type distinctions across the entire numeric hierarchy, including complex numbers.
+- Added strict type checking support for nested and union types (e.g., `list[int]`, `dict[float, int]`, `Union[int, float]`)
 
 ### 🔧 Internal changes